# Optimize jsonObfuscator using sync.Pool and smart buffers allocation