--- a/pkg/obfuscate/json.go
+++ b/pkg/obfuscate/json.go
@@ -168,8 +168,8 @@ func (p *jsonObfuscator) obfuscate(data []byte) (string, error) {
 
 		case scanObjectKey:
 			// done scanning key
-			k := strings.Trim(string(keyBuf), `"`)
-			if !st.keeping && p.keepKeys[k] {
+			k := string(bytes.Trim(keyBuf, `"`))
+			if !st.hasFlag(keepingFlag) && p.keepKeys[k] {
 				// we should not obfuscate values of this key
 				st.keeping = true
 				st.keepDepth = depth + 1