--- a/pkg/obfuscate/json.go
+++ b/pkg/obfuscate/json.go
@@ -6,6 +6,7 @@
 package obfuscate
 
 import (
+	"bytes"
 	"strconv"
 	"strings"
 )