--- a/releasenotes/notes/faster-obfuscation-744c6710d52ad6d9.yaml
+++ b/releasenotes/notes/faster-obfuscation-744c6710d52ad6d9.yaml
@@ -1,0 +1,7 @@
+---
+enhancements:
+  - |
+    Improved performance and memory consumption, both halved on average.
+fixes:
+  - |
+    Concurrency issue at high volumes fixed.