### What does this PR do?

Improves from #19643 fixed implementation. The fix caused an increment of allocations, up to 13 per op.

The optimizations reduces allocations to 2 per op, 3 per op in the worst case.

There is still some room for optimization, but the code would be hard to read and maintain.

### Motivation

I measured the impact of allocating the new internal state struct and I found that it was significant (in bytes and nanoseconds scale):

main:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1381185               855.3 ns/op           384 B/op          7 allocs/op
```

This PR:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1000000              1053 ns/op             576 B/op         13 allocs/op
BenchmarkObfuscateJSON/obfuscate.mongo.json.keep_values-10                               1979707               606.8 ns/op           400 B/op         11 allocs/op
```

I applied two optimizations ([`out.Grow()`](https://github.com/DataDog/datadog-agent/pull/19643/files#diff-7ce08f4ef3a4c33487f631696e3fe94835170c85781693113325c558be655abbR109) and [`bytes.Trim()` instead of `strings.Trim()`](https://github.com/DataDog/datadog-agent/pull/19643/files#diff-7ce08f4ef3a4c33487f631696e3fe94835170c85781693113325c558be655abbR172)) that improved performance, setting it closer to the original implementation:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1268534               943.5 ns/op           440 B/op          9 allocs/op
BenchmarkObfuscateJSON/obfuscate.mongo.json.keep_values-10                               2123896               549.2 ns/op           312 B/op          7 allocs/op
```

Introducing `sync.Pool` allocations are down to 1-3 allocs/op:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                    	 1505170	       798.2 ns/op	     224 B/op	       2 allocs/op
BenchmarkObfuscateJSON/obfuscate.mongo.json.keep_values-10                        	 2821915	       424.2 ns/op	     160 B/op	       2 allocs/op
```

PS: if we apply the optimizations to main, except `sync.Pool`, we see that the performance impact of the fix would be proportionally to the baseline, doubling allocation. At least it has been mitigated and set almost to the original implementation's performance.

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1585716               754.8 ns/op           248 B/op          3 allocs/op
```

### Describe how to test/QA your changes

Unit tests are in place to test and benchmark the implementation.

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.